requirements.txt
streamlit
pandas
pyyaml
pillow
faiss-cpu
transformers
torch
plotly
numpy

